feat: service간 의존 제거
